backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
aiohttp==3.13.2
openai>=1.55.3
httpx>=0.28,<0.29
httpcore>=1.0.0
python-multipart==0.0.9
